Using dockerfiles and cloud services/requirements.txt
dash
pandas
datetime
plotly
requests
s3fs
fsspec
python-dateutil